Float values for position cost

+++ EventsApp/models.py
@@ -251,7 +251,7 @@ class Events_PositionInfo(models.Model):
     max_age = models.IntegerField(blank=True, null=True)
     min_age = models.IntegerField(blank=True, null=True)
     no_of_position = models.IntegerField(blank=True, null=True)
-    position_cost = models.IntegerField(blank=True, null=True)
+    position_cost = models.FloatField(blank=True, null=True)
 
     def __str__(self):
         return self.event.event_title + " Position_" + str(self.position_number)
